Move healthcheck stuff out
