+++ src/rust/rust-proto-new/src/graplinc/grapl/api/pipeline_ingress/v1beta1.rs
@@ -1,3 +1,4 @@
+#![allow(warnings)]
 use bytes::{
     Bytes,
     BytesMut,
@@ -149,28 +150,6 @@ impl SerDe for PublishRawLogResponse {
 }
 
 //
-// gRPC
-//
-
-#[non_exhaustive]
-#[derive(Debug, Error)]
-pub enum HealthcheckError {
-    #[error("not found {0}")]
-    NotFound(String),
-
-    #[error("healthcheck failed {0}")]
-    HealthcheckFailed(String),
-}
-
-#[non_exhaustive]
-#[derive(Debug)]
-pub enum HealthcheckStatus {
-    Serving,
-    NotServing,
-    Unknown,
-}
-
-//
 // client
 //
 
@@ -188,24 +167,12 @@ pub mod client {
         protobufs::graplinc::grapl::api::pipeline_ingress::v1beta1::pipeline_ingress_service_client::PipelineIngressServiceClient as PipelineIngressServiceClientProto,
         SerDeError,
     };
+    use crate::protocol::healthcheck::{HealthcheckError, HealthcheckStatus};
 
     use futures::{FutureExt, TryFutureExt};
     use thiserror::Error;
     use tokio::time::error::Elapsed;
     use tonic::Request;
-    use tonic_health::{
-        proto::{
-            health_client::HealthClient as HealthClientProto,
-            health_check_response::ServingStatus as ServingStatusProto,
-            HealthCheckRequest as HealthCheckRequestProto
-        },
-    };
-
-    use super::{
-        HealthcheckError,
-        HealthcheckStatus
-    };
-
     #[non_exhaustive]
     #[derive(Debug, Error)]
     pub enum ConfigurationError {
@@ -259,115 +226,6 @@ pub mod client {
                 .await
         }
     }
-
-    pub struct HealthcheckClient {
-        proto_client: HealthClientProto<tonic::transport::Channel>,
-        service_name: &'static str,
-    }
-
-    impl HealthcheckClient {
-        #[tracing::instrument]
-        pub async fn connect<T>(
-            endpoint: T,
-            service_name: &'static str,
-        ) -> Result<Self, ConfigurationError>
-        where
-            T: std::convert::TryInto<tonic::transport::Endpoint> + std::fmt::Debug,
-            T::Error: std::error::Error + Send + Sync + 'static,
-        {
-            Ok(HealthcheckClient {
-                proto_client: HealthClientProto::connect(endpoint).await?,
-                service_name,
-            })
-        }
-
-        #[tracing::instrument(skip(self))]
-        pub async fn check_health(&mut self) -> Result<HealthcheckStatus, HealthcheckError> {
-            let request = HealthCheckRequestProto {
-                service: self.service_name.to_string(),
-            };
-
-            let response = match self.proto_client.check(request).await {
-                Ok(response) => response.into_inner(),
-                Err(e) => match e.code() {
-                    tonic::Code::NotFound => return Err(HealthcheckError::NotFound(e.to_string())),
-                    _ => return Err(HealthcheckError::HealthcheckFailed(e.to_string())),
-                },
-            };
-
-            match response.status() {
-                ServingStatusProto::Serving => Ok(HealthcheckStatus::Serving),
-                ServingStatusProto::NotServing => Ok(HealthcheckStatus::NotServing),
-                ServingStatusProto::Unknown => Ok(HealthcheckStatus::Unknown),
-                ServingStatusProto::ServiceUnknown => Err(HealthcheckError::HealthcheckFailed(
-                    "service unknown".to_string(),
-                )),
-            }
-        }
-
-        #[tracing::instrument]
-        pub async fn wait_until_healthy<T>(
-            endpoint: T,
-            service_name: &'static str,
-            timeout: Duration,
-            polling_interval: Duration,
-        ) -> Result<Self, ConfigurationError>
-        where
-            T: std::convert::TryInto<tonic::transport::Endpoint> + Clone + std::fmt::Debug,
-            T::Error: std::error::Error + Send + Sync + 'static,
-        {
-            let client_fut = async move {
-                let mut healthcheck_client = loop {
-                    match HealthcheckClient::connect(endpoint.clone(), service_name).await {
-                        Ok(client) => break client,
-                        Err(e) => {
-                            tracing::warn!(
-                                "could not construct healthcheck client for {}, waiting {}ms: {}",
-                                service_name,
-                                polling_interval.as_millis(),
-                                e,
-                            );
-                            tokio::time::sleep(polling_interval).await;
-                        }
-                    }
-                };
-
-                loop {
-                    match healthcheck_client.check_health().await {
-                        Ok(result) => match result {
-                            HealthcheckStatus::Serving => {
-                                tracing::info!("{} is serving requests", service_name);
-                                break Ok(healthcheck_client);
-                            }
-                            other => {
-                                tracing::warn!(
-                                    "{} is not yet serving requests, waiting {}ms: {:?}",
-                                    service_name,
-                                    polling_interval.as_millis(),
-                                    other
-                                );
-                                tokio::time::sleep(polling_interval).await;
-                            }
-                        },
-                        Err(e) => match e {
-                            HealthcheckError::HealthcheckFailed(_) => break Err(e),
-                            HealthcheckError::NotFound(reason) => {
-                                tracing::warn!(
-                                    "{} healthcheck not found yet, waiting {}ms: {}",
-                                    service_name,
-                                    polling_interval.as_millis(),
-                                    reason
-                                );
-                                tokio::time::sleep(polling_interval).await;
-                            }
-                        },
-                    }
-                }
-            };
-
-            tokio::time::timeout(timeout, client_fut.map_err(|e| e.into())).await?
-        }
-    }
 }
 
 //
@@ -405,10 +263,6 @@ pub mod server {
         Server,
     };
 
-    use super::{
-        HealthcheckError,
-        HealthcheckStatus,
-    };
     use crate::{
         graplinc::grapl::api::pipeline_ingress::v1beta1::{
             PublishRawLogRequest,
@@ -421,6 +275,11 @@ pub mod server {
             },
             PublishRawLogRequest as PublishRawLogRequestProto,
             PublishRawLogResponse as PublishRawLogResponseProto,
+        },
+        protocol::healthcheck::{
+            server::init_health_service,
+            HealthcheckError,
+            HealthcheckStatus,
         },
         SerDeError,
     };
@@ -513,7 +372,7 @@ pub mod server {
         T: PipelineIngressApi<E> + Send + Sync + 'static,
         E: ToString + Send + Sync + 'static,
         H: Fn() -> F + Send + Sync + 'static,
-        F: Future<Output = Result<HealthcheckStatus, HealthcheckError>> + Send,
+        F: Future<Output = Result<HealthcheckStatus, HealthcheckError>> + Send + Sync + 'static,
     {
         api_server: T,
         healthcheck: H,
@@ -530,7 +389,7 @@ pub mod server {
         T: PipelineIngressApi<E> + Send + Sync + 'static,
         E: ToString + Send + Sync + 'static,
         H: Fn() -> F + Send + Sync + 'static,
-        F: Future<Output = Result<HealthcheckStatus, HealthcheckError>> + Send,
+        F: Future<Output = Result<HealthcheckStatus, HealthcheckError>> + Send + Sync + 'static,
     {
         /// Construct a new gRPC server which will serve the given API
         /// implementation on the given socket address. Server is constructed in
@@ -570,67 +429,14 @@ pub mod server {
         /// Run the gRPC server and serve the API on this server's socket
         /// address. Returns a ConfigurationError if the gRPC server cannot run.
         pub async fn serve(self) -> Result<(), ConfigurationError> {
-            let (mut health_reporter, health_service) = tonic_health::server::health_reporter();
-
-            // we configure our health reporter initially in the not_serving
-            // state s.t. clients which are waiting for this service to start
-            // can wait for the state change to the serving state
-            health_reporter
-                .set_not_serving::<PipelineIngressServiceServerProto<PipelineIngressProto<T, E>>>()
-                .await;
-
-            let healthcheck_handle = tokio::task::spawn(async move {
-                // I initially tried to break this loop out into its own
-                // function, but I ran into this issue:
-                //
-                // https://github.com/rust-lang/rust/issues/83701
-                //
-                // Unfortunately, the need to parametrize such a function by
-                // self.healthcheck's type and the parameter S in
-                // HealthReporter::set_serving<S>(..) makes this awkward, so I
-                // just inlined the whole thing here.
-                loop {
-                    match (self.healthcheck)().await {
-                        Ok(status) => match status {
-                            HealthcheckStatus::Serving => {
-                                tracing::info!("healthcheck status \"serving\"");
-                                health_reporter
-                                        .set_serving::<PipelineIngressServiceServerProto<
-                                            PipelineIngressProto<T, E>,
-                                        >>()
-                                        .await
-                            }
-                            HealthcheckStatus::NotServing => {
-                                tracing::warn!("healthcheck status \"not serving\"");
-                                health_reporter
-                                        .set_not_serving::<PipelineIngressServiceServerProto<
-                                            PipelineIngressProto<T, E>,
-                                        >>()
-                                        .await
-                            }
-                            HealthcheckStatus::Unknown => {
-                                tracing::warn!("healthcheck status \"unknown\"");
-                                health_reporter
-                                        .set_not_serving::<PipelineIngressServiceServerProto<
-                                            PipelineIngressProto<T, E>,
-                                        >>()
-                                        .await
-                            }
-                        },
-                        Err(e) => {
-                            // healthcheck failed, so we'll set_not_serving()
-                            tracing::error!("healthcheck error {}", e);
-                            health_reporter
-                                .set_not_serving::<PipelineIngressServiceServerProto<PipelineIngressProto<T, E>>>()
-                                .await
-                        }
-                    }
-
-                    tokio::time::sleep(self.healthcheck_polling_interval).await;
-                }
-            });
-
             // TODO: add tower tracing, tls_config, concurrency limits
+            let (healthcheck_handle, health_service) =
+                init_health_service::<
+                    PipelineIngressServiceServerProto<PipelineIngressProto<T, E>>,
+                    _,
+                    _,
+                >(self.healthcheck, self.healthcheck_polling_interval)
+                .await;
             Ok(Server::builder()
                 .add_service(health_service)
                 .add_service(PipelineIngressServiceServerProto::new(
